Save to PNG with simplified API
